requirements.txt
hydra-core==1.3.1
librosa==0.9.2
numba==0.56.4
numpy==1.22.4
pytorch-lightning==1.8.6
resampy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
tensorboardX==2.5.1
tqdm==4.64.1
